refactor: ♻️ move partials projects files
