+++ components/partials/create/project/parts/CreateProjectFields.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "next-i18next";
 import { CreateProjectValues } from "../CreateProjectForm";
 
 // Steps
-import { getSectionsByProjectType } from "components/partials/project/projectSections";
+import { getSectionsByProjectType } from "components/partials/project/[id]/projectSections";
 
 // Components
 import { Stack } from "@bbtgnn/polaris-interfacer";
